vision-bench/requirements.txt
pyyaml>=6.0
openai>=1.0
anthropic>=0.40
mistralai>=1.0
